requirements.txt
# Core dependencies for OpenUQFOAM
numpy>=1.20.0
pandas>=1.3.0
jinja2>=3.0.0
pyyaml>=6.0.0
matplotlib>=3.5.0
tqdm>=4.60.0

# Optional dependencies
pyDOE2>=1.3.0  # For improved Latin Hypercube Sampling
scipy>=1.7.0   # For statistical functions